requirements.txt
flask~=2.0.1
requests~=2.26.0
terra-python~=0.0.11
python-dotenv~=0.19.0